textagram_ai/requirements.txt
anyio==3.6.2
black==22.10.0
brotlipy==0.7.0
certifi==2022.9.14
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_62rp5d8fd4/croots/recipe/cffi_1659598655556/work
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_3evwafgyg8/croot/cryptography_1665612651044/work
customtkinter==4.6.1
darkdetect==0.7.1
diffusers==0.3.0
fastapi==0.87.0
filelock==3.8.0
gcloud==0.18.3
googleapis-common-protos==1.57.0
h11==0.14.0
httplib2==0.21.0
huggingface-hub==0.9.1
idna @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_00jf0h4zbt/croot/idna_1666125573348/work
importlib-metadata==4.12.0
jws==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/0cda23d8-7460-44b2-7e5d-3c76a8a0ca7e/volume/mkl_random_1626186083266/work
mkl-service==2.4.0
mypy-extensions==0.4.3
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cd2y6umv2d/croot/numpy_and_numpy_base_1668593733683/work
oauth2client==4.1.3
packaging==21.3
pathspec==0.10.2
Pillow==9.2.0
platformdirs==2.5.4
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.15.0
pydantic==1.10.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
Pyrebase4==4.5.0
PySocks @ file:///opt/concourse/worker/volumes/live/112288ac-9cb0-4e73-768b-13baf4ca6419/volume/pysocks_1605305820043/work
python-jwt==2.0.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
starlette==0.21.0
tk==0.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.22.1
typing_extensions @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ff5_5nqr6l/croots/recipe/typing_extensions_1659638832447/work
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_7f7kb5tudl/croot/urllib3_1666298941688/work
uvicorn==0.19.0
zipp==3.8.1
google-cloud-logging